+++ packages/nextra/__test__/fixture/page-maps/respect-order-for-type-separator-menu-and-item-with-href/one/two/_meta.ts
@@ -1,0 +1,19 @@
+export default {
+  menu: {
+    type: 'menu',
+    items: {
+      nextra: {
+        title: 'Nextra',
+        href: 'https://nextra.site'
+      }
+    }
+  },
+  '---': {
+    type: 'separator'
+  },
+  qux: '',
+  nextra: {
+    title: 'Nextra',
+    href: 'https://nextra.site'
+  }
+}